Removed detector_label() from visible module functions. Renamed to make_detector_label(). Moved make_saturation_label() from gen_tso_app.py to pandeia_defaults.py Added order argument to both functions.
